requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.18.0
requests>=2.31.0
numpy>=1.26.0
matplotlib>=3.8.0
wordcloud>=1.9.3
kiwipiepy>=0.17.1
python-dotenv>=1.0.0
